Refactor parameter definitions to integrate new variation buttons for all parm types
- Added ScalarNewVariationButtons to BooleanParamDefinition and NumberParamDefinition for enhanced control.
- Updated PaletteParamDefinition and Vector4ParamDefinition to include respective new variation buttons.
- Improved ParamDefinition structure to support rendering of new variation buttons based on parameter type.
