+++ src/paramDefinitions/NumberParam.ts
@@ -2,6 +2,7 @@ import { VJScalarParameterControl } from "@/src/components/VJPage/VJScalarParame
 import { ScalarParameterControl } from "@/src/components/PatternPlayground/ScalarParameterControl";
 import { isNumberParam } from "@/src/types/PatternParams";
 import { createParamType } from "@/src/paramDefinitions/ParamDefinition";
+import { ScalarNewVariationButtons } from "@/src/components/ParameterVariations/ScalarNewVariationButtons";
 
 // explicit nominal typing would be better
 
@@ -9,4 +10,5 @@ export const NumberParamDefinition = createParamType<number>({
   isParamType: isNumberParam,
   ParameterControl: ScalarParameterControl,
   VJParameterControl: VJScalarParameterControl,
+  NewVariationButtons: ScalarNewVariationButtons,
 });
